pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "easeai"
version = "0.1.0"
description = "Reusable XGBoost + SHAP workflow for tabular regression explainability"
readme = "README.md"
requires-python = ">=3.9"
license = {text = "MIT"}
authors = [
  {name = "Ria A. Martins"}
]
dependencies = [
  "numpy>=1.24",
  "pandas>=2.0",
  "scikit-learn>=1.3",
  "xgboost>=2.0",
  "shap>=0.45",
  "matplotlib>=3.8",
  "Pillow>=10.0",
]
keywords = ["xgboost", "shap", "tabular", "xai", "regression"]
classifiers = [
  "Programming Language :: Python :: 3",
  "License :: OSI Approved :: MIT License",
  "Operating System :: OS Independent",
  "Intended Audience :: Science/Research",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0",
]

[tool.setuptools]
include-package-data = true

[tool.setuptools.packages.find]
include = ["easeai*"]
